backend/requirements.txt
# Core Framework - Python 3.8 compatible
fastapi==0.95.2
uvicorn==0.21.0
pydantic==1.10.12
python-multipart==0.0.6

# Database
sqlalchemy==1.4.46
psycopg2-binary==2.9.9
alembic==1.12.1
asyncpg==0.28.0

# Async
aiofiles==23.2.1
aiohttp==3.8.1

# Security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
cryptography==37.0.4
PyJWT==2.6.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Testing
pytest==7.4.0
pytest-asyncio==0.21.1
httpx==0.24.0

# Logging
python-json-logger==2.0.7

